pyproject.toml
[project]
name = "word-mcp-server"
version = "0.1.0"
description = "Word document MCP server — fills .docx templates and manages versioned outputs via SharePoint and OneDrive"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "starlette>=0.46.1",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.0.0",
    "python-docx>=1.1.2",
    "docxtpl>=0.20.2",
    "msal>=1.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"